size_t: use for type consistency where necessary

+++ libsrc/v1hpg.c
@@ -1452,7 +1452,7 @@ nc_get_NC(NC3_INFO* ncp)
 	      		if(extent > 4096)
 				extent = 4096;
 			if(extent > filesize)
-			        extent = filesize;
+			        extent = (size_t)filesize;
 		}
 		else if(extent > ncp->chunk)
 		    extent = ncp->chunk;
